made it so that the new LiveWindow mess doesn't do anything
Calls to LiveWindow (which is a kind of front end for smart dashboard) got added
to all of the low-level classes (aka the ones that we use). We decided against
creating a fork of WPILib for all of the parts that we care about and instead
just made it so that all of the function calls don't do anything.

git-svn-id: https://robotics.mvla.net/svn/frc971/2013/trunk/src@4179 f308d9b7-e957-4cde-b6ac-9a88185e7312