added a lifting timeout
Sometimes, a frisbee doesn't get all of the way into the loader which
means that it never gets to the top. The way that this problem used to
get fixed was the loader would come down, release, it'd try to load the
next frisbee, and it would push it all the way in, leaving a frisbee in
the loader at the end. With the sensor, it just gets stuck trying to
lift, so I added a timeout to fix that.