updated gyp to a newer version that doesn't import itself from sys.path
The old version of gyp would import (all of) itself from sys.path if it
could, which caused problems with people who had local installations. In
order to work with this new version, I also had to update the version of
ninja that was used and the way that the cRIO setup creates .out files.