fixed frc971/constants and all the uses of it
Before, it used aos::robot_state to get the team number, which meant a
lot of complication around dealing with that and the fact that it's not
there at the very beginning. Now, it just uses the IP address to do it,
which means that the interface is way simplier and nicer.