I updated the claw as best as I could to have independent separation/position. It won't work this way.
Currently, I had to make some simplifications in the model in order to be able to manipulate all 4 eigenvalues of the A - B * K matrix in the claw and those simplifications make the model inaccurate and therefore it doesn't actually properly control things and it just doesn't work. I am committing this for the sake of a record, but this should not atually be used.