implement robust mutex support
This allows making everything in shared memory robust to
processes dying at any point in time (not actually done yet in this
commit).

This includes using FUTEX_REQUEUE_PI, but currently only on ARM because
that's the only place we can rely on the kernel not corrupting random
memory due to a bug (fix has been merged upstream).

Change-Id: Id5bda1dc3185a1aac759510934bce6fd9121ad3f