sensor_sim: Add documentation and clear up some issues.
The sensor simulator turned somewhat into a mess, largely because of
two reasons:
- Many people disagreeing on what parameters should mean etc.
- No tests to verify the absence of bugs.

This commit aims to fix that by cleaning up some of the code, adding
a bunch of documentation and most importantly by adding a test for
the class.

The `position_sensor_sim.cc' file now contains an explanation along
with an ASCII diagram of what problem the class is meant to solve.

Change-Id: I34d203976ceefead50096ddc0d0fcfef12af6d87