Add seasocks to aos externals for http_status and add test server.
This is stage 1 of the http_status project!
Seasocks requires a rather elaborate method of compiling the code
and including the web pages into the compiled executable as raw data,
so some of the stuff in the gyp system may not be pretty. We will fix
these problems up in later revisions as needed.

The web directory holds the rudimentary files for the web server (404
message, javascript library that we will use in later stages of
designing th web page interface, and icons. web_test holds the files for
the test webpage that can be accessed at localhost:8080 when the server
is running. Again, this directory, along with the content of
http_status.cc, is temporary and intended only to show the functionality
of the server socket.

Change-Id: I2ad1598ae200b526620b014db970e980f3c8a563