Make preliminary Seasocks server with data collector.
This is stage 2 of the http_status project!
Currently, this version can display various things from the queues in
a web browser, but only in plain text. It implements some of the basic
classes that will later be used for later stages. It also supports
concurrent connections.

A couple of things not included are all the trigger/bandwidth adjust
options.

Change-Id: I108d965a768c073a0f4b3757c0475fe7bfc34ccb