Add elevator control loop and test.
This commit includes a big chunk of elevator code that Comran worked on,
and all the zeroing code that Adam got working for the bot3.
Unfortunately, we did not split these two projects apart earlier, so
this is a big commit containing all of the code combined. Right now, all
of the elevator tests that aren't commented out pass with the hall
effect zeroing method that Adam made, but all the tests in
elevator_lib_test will need to be checked by other programmers on the
team & updated as needed in a future commit.

Change-Id: I4ad6247303c3a939a28eff7d2043de74a2d3865d