Fix a drivetrain polytope bug
When both lines intersect the box, the code was assuming that the
intersection was farther from the origin than the box. However, that's
not true when the drivetrain has a long way to drive and a small angle
error.

This fixes the original issue which led to the polytope code for the
drivetrain being disabled, but there are other numerical precision
issues that are now showing up, so it's still not enabled.

Change-Id: I99a1299e538a395685c2e7e555dfccab48055349