Work on getting wpilib_interface finished
I moved all the constants together to make it somewhat easier to work. I
also finished renaming serializer to indexer because Austin is more
stubborn than the rest of the team combined.

It at least compiles now. The TODOs are also also much more specific to
the few remaining pieces.

Change-Id: Id3baf2b5be0a946345b152612c2099cf1b599727