Remove some mostly-dead code dealing with the flash controller
It did weird things with the clocking in order to talk to the flash
controller to read some config data (maybe one of the one-time-writable
segments?). We never fill this out in the first place, so just remove it
to simplify. If we ever want unique serial numbers, the chips we care
about have an actual unique ID.

Change-Id: I4a214f3c49232081b67c896b8c57a5d1c3b8e69c