Add a Bazel-built slicot via f2c
Previously, we were using the host binaries along with slycot, which
didn't support using it at runtime.

We built slicot itself using f2c, and then the normal Bazel C/C++
infrastructure on the resulting C code.

We're also using clapack, which is lapack and BLAS pre-run through f2c
along with various fixups.

The build is horribly noisy and full of warnings and information
messages, but oh well.

Change-Id: If13056ff56d334b7207cd9f0f7c0ce22b999b78e