Added CG offset from center to physics model.
Updates the physics to model the effect of the CG located to the left
or right of the center of the robot.

It would be good to add a test case to show that the changes work.
Please help me with this by suggesting how I might add a test
case.

Change-Id: I19c18209cd380293d7d35a52d7214bbfef4acfd6