Sanify the MCU linker scripts
We now have SRAM_L and SRAM_U split apart. Otherwise, it's just a matter
of time before the linker splits something across them which the
compiler tries accessing with ldrd or something and causes mysterious
problems at runtime.

Also, the new arrangement has the stack grow off the end of RAM, so
instead of clobbering the heap you'll just get a hard fault if you
overflow the stack.

Change-Id: I3c18ac665b48ddc0456e0a1ef34213dcac8e9c85