Support varying table sizes in the fast math code
Our code is currently running pretty tight on memory on the K22 (used on
fet12v2). Our encoder there also doesn't as many counts as the ones on
the pistol grip. Therefore, with a bit of cleverness and templates, we
can free up some RAM to have enough breathing room to split up the heap
and stack better.

Change-Id: Ibea73887b5570326ec1338b296f71ecab3d4ba7a