Convert CROSSTOOL file mostly over to features
This is in preparation for the new compiler for 2019.  We need to move a
flag in a way that isn't possible with the legacy fields.

bazel clean --expunge && bazel build -c opt --cpu=roborio //...
--experimental_execution_log_file=/tmp/exec_new.log &&
../bazel/bazel-bin/src/tools/execlog/parser --log_path=/tmp/exec_new.log
--log_path=/tmp/exec_orig.log --output_path=/tmp/exec2.log.txt
--output_path=/tmp/exec1.log.txt && cat /tmp/exec2.log.txt | grep -v
'runner: .*$' | grep -v 'remote_cache_hit: .*$' >
/tmp/exec2_filtered.log.txt && cat /tmp/exec1.log.txt | grep -v 'runner:
.*$' | grep -v 'remote_cache_hit: .*$' > /tmp/exec1_filtered.log.txt

(/tmp/exec_orig.log was created from a run before any changes)

Change-Id: I475ed6baba8d93aba8c0c9ecbf60c4dd7d13a2cb