Scripts for backing up and restoring ODROID microSD cards
The backup_ODROID_microSD_card.rb and
restore_ODRIOD_microSD_card_from_backup.rb scripts are used to backup
ODROID microSD cards and create new ones.  They use tar, mkfs, and dd.
dd is only used for the master boot record.  The traditional way of
copying a mircoSD card is to use dd to make a copy of the entire disk
and then copy that back using dd.  This can take a long time - hours -
and requires having enough space to save a file that is the size of the
microSD card.  These scripts run in less than 5 minutes each and use
only as much disk space as is require to hold the compressed contents
of the microSD card.

Change-Id: I763bcf8235596cc3e7dbba3ab3ec35fd6c680d5d