Compute spline derivatives as a function of distance
We want to do all our math as a function of the distance along the path,
not the bezier spline parameter.  So, take all the derivatives
accordingly.  This also requires us to be able to compute the path
length accurately and convert from path length to the spline parameter.

Next step is to start planning the max velocity we can follow the spline
at to adhere to our constraints.

Change-Id: I7c6d9ddd2ce92aa22e334eaa02594b91ac11066a