Store the strategy used for each spline segment
We were running into an issue where the ff voltage would chatter.  The
core problem was that we planned a segment using one acceleration
function, and executed it with another.  That's not consistent with the
velocity plan, whether or not it's right.  We need the acceleration that
we use to execute the plan to be consistent with the velocity, which
only works if we use the same acceleration function used to integrate up
the velocity during the segment.  It's more important to be internally
consistent than to track constraints perfectly.

Change-Id: Id28c7036c2d3e585a99577cb4e03967f2bba3e26