Analyse robot log file to understand stilts lift
The robot lift time was much longer than expected when it was
first tested.  This script reads the log file and creates plot
files and a gnuplot script for viewing the results.  From
viewing these results, it was determined that the lift
control loop profile acceleration needed to be increased
from 0.5 m/s^2 to 2.0 m/s^2.

Change-Id: I69ef6ea36327301d581918d59436a0fb2acb8ee9