Increase stilts lift profile velocity goal
From looking at the log files, it appears that the robot might
be able to raise itself up a bit faster than it currently is doing.
I have increased the goal max velocity from 0.45 to 0.65 m/s.
When Austin tuned the acceleration goal, he increased it from 0.5 m/s^2
to 2.0 m/s^2.  I did the same in this change.

Change-Id: I88cdce5cb9dec467d5ff7a8d6074b7fee69355c2