Optimize discretization of A/Q for EKF
Because we are recomputing the discretization of A/Q for every single
step in the EKF, it had been taking up ~70-80% of the total computation
time of the entire EKF. This knocks it down to something like 50% of the
total EKF computation time (for ~3-4x speed improvements overall).

Because the discretizations of A/Q are dependent on both dt and on the
heading of the robot, they are not trivially cacheable. Future
improvements would be:
-Only recalculate discretization when the timestep changes and assume
 that the heading estimate will only change by minute amounts and
 so we don't need to redescritize unless dt has changed.
-Figure out how to cache 7x7 components of each that should stay
 constant for constant dt.
-Determine whether any convenient relationships hold with respect to
 dt/heading that would let us make a reasonably small interpolation
 table.

Change-Id: I8838d4837f380a2411ab9da99d542cdbe9999e41