Add timing reports for both simulation and shm
This involves a major refactor of how event loops are put together.
Lots of functionality for timing reports is in the base event loop
class (though hidden as much as possible to avoid cluttering the
interface).  More functionality is now in the base class, with simpler
and more consistent subclasses.

Change-Id: I163ffb8d1b0e29a0231b54e205c2d7c5241d662b