Sort wakeups and add send latency.
We don't want to rely on epoll to trigger wakeups in order.  Timers will
likely wake up before watchers due to less kernel delays, etc.  So,
maintain a sorted list of events and see which should be triggered first
and trigger it first.  It also merges wakeups nicely.

This gives us the infrastructure to handle the same problem in
simulation.  We can have perfect timers, and senders with latency.  But,
if a timer wakes up while something is being sent, we will move up the
wakeup and handle it in order.

Change-Id: I8675ec5221dd2603b4aa5b1a3729907a599813b4