Update Phoenix libraries for 2020
With this update, the phoenix diagnostics server now runs all the time
in the wpilib_interface binary. Not sure if this is a good thing or a
bad thing, but having it enabled is the only way to update firmware or
do any configuration of CTRE devices with the new 2020 CAN stack. There
may be some flag that llows us to disable it.

Also, adds "-ldl" to the ni libraries because somehow updating the
Phoenix libraries caused us to start needing this.

Change-Id: Ifd7b2b108621058e51fa9d4eb4f16f1903a42337