Handle a large number of messages being published
writev has a limit for the number of iovec's that can be written at
once.  If this is exceeded, it fails to write data.  This is an
independent check from us wanting to flush if there has been too big a
quantity of data written.  A burst of small messages can trigger this.

Check for this and flush when we get to the limit.

Change-Id: Ie8096f5ae734de9f6716d2ba1846c42f6c80ff3b