Handle log files not starting at the same time.
The monotonic clocks were assumed to be in sync.  That isn't realistic.
This assumption leaked into how we kept the queues primed, and how the
event loop was initialized.

This isn't enough to actually replay in sync.  We are assuming that the
realtime clocks are in sync and the monotonic clocks don't drift from
each other.  That'll be good enough to get started, but not for long.

Change-Id: Ic18e31598f1a76edee0b0d5a2d7936deee1fbfec