Fix action_test to not require watchers to watch while not running
When we are not in Run/RunFor, we aren't technically running, so
watchers shouldn't queue up events.  This is technically correct, but a
bit annoying.  We either need to track if event loops have run, or if
they are running now.  It is easier to track if they are running now.

Fix the action_test to not require this behavior so when the
implementation changes, it will still work.

Change-Id: I689244463b98be0c210bb6adc223bcff21c277d2