Make logger and timesamp_logger config lists
We want to log messages on both the roboRIO, and on the laptop/logger
pi.  This means that we want lists of loggers instead of a single
logger.

We are lucky and there are no known log files with loggers in them which
we care to preserve.  This lets us reuse the field without worying about
compatability.

Change-Id: I452994fbca78909d6320da3d02ad51fe45a0354b