Path-Relative LQR for trajectory following
Implement a basic LQR controller that operates on a path-relative state
for spline control. This formulation also helps to leave the path open
for changes around how we manage gains (e.g., setting different cost
matrices at different points along the path) as well as leaving room
to move into a more MPC-like formulation where we more explicitly
compensate for saturation.

Left the tuning a bit loose so far, since the control
scheme implicitly assumes that saturation is not an issue.

Change-Id: I792bc939735b405b09ba4b8af777a1b2b242d325