Basic Typescript WebGL Plotter
This implements a basic plotter in typescript using WebGL.

Based on my cursory benchmarks, a reasonable computer within any crazy
graphics capabilities should be able to plot ~1-3 million points using
the plotting tool without major issue (the included sample plots 2
million points).

This also creates a sample webpage that uses the typescript reflection
library to subscribe to the timing report generated by the web proxy
application and then plots some of the statistics in realtime.

This does not yet support useful logfile analysis, because the web proxy
does not currently provide useful support for reading from a logfile
(technically it could do so as-is, but doesn't provide any good way to
ensure that you can get all the messages from the logfile).

Change-Id: Ibfe709259f3172530d94e2ad1aa4f257db7be9e0