Automatically copy unmodified fields in MergeConfiguration
We weren't copying channel_storage_duration, and the previous setup
meant that if we were to add a new top-level field to the Configuration
flatbuffer it wouldn't automatically get merged. This does have a slight
performance impact due to the excessive copying, but I don't think
anyone cares all that much about the performance of MergeConfiguration.

Change-Id: I3e4e11a052ed53db8dc374b037a88627b0e5e938