Clean up the basic codelab
The config was incomplete because it didn't include timing reports, and
wasn't even using the standard config process anyways. The comments were
also out of date due to the flatbuffers transition, and I added in a bit
more boilerplate in the RunIteration function, since handling the
flatbuffer builders is a bit less trivial than the queues were.

Change-Id: I6a9b12ca6c925b561211937fcf36d932299946d2