Add UUIDs for both the logger and parts
These aren't required to be used, but they help track which parts should
go together, and which parts came from the same logger.  The former
being more interesting than the latter, but if we are here, why not
track both?

Follow-up review will add a "sort bunch of files into corresponding
parts" function.

Change-Id: I1e17ae8617b267bc6c5d90ff9d7feb351c0c4b47