Support node specific remaps in LogReader
When we have a bunch of nodes running similar pieces of software, we end
up with a bunch of channels which are all the same, bar a prefix.  The
applications then end up with node specific maps to avoid being
different everywhere.

When renaming those in replay, we need to recreate both the node
specific map, and take the map into account when renaming.  We don't
want to leak the naming pattern up.

Change-Id: I6a6a0556bf49664679f443a21eaed23d024a515f