Teach aos_dump that not all channels are readable everywhere
When we have dozens of channels on a dozens nodes, we have a bunch of
irrelevant channels being displayed.  This makes it hard.  Instead,
teach aos_dump to only display the channels that the user can show (with
a flag to override in case the user actually wants something else).

Change-Id: Ib0a7a9303fa1862aa622d34e309374e9e519429e