Recreate remote timestamp logging in LogReader
It is useful to be able to log data, replay it into a simulation, and
then recreate a log again.  To do this, we need remote timestamps to
work correctly.

When LogReader replays a forwarded message, it now creates the
corresponding MessageHeader and publishes it.  It also tracks the queue
indicies such that the message is valid and can be logged.

Logger also translates channel indices as well when the logging config
is not the event loop config.

Change-Id: Iff6175a204b191c6f43a1d73ffce5b542925860c