Handle only some nodes having data
We used to require either all nodes to communicate, or none.  Obviously,
when you have a bunch of nodes, this starts to fall down.  Instead,
solve for the nodes we have relative timestamp information for, and
assume the rest match the distributed clock.

This is useful when a <-> b <-> c <-> d and you only have logs from a,
or a and b.

Change-Id: I883b7da03a0226f27c3b07b0590585b642ecb671