Explicit ids, and make nodes a list in Application
Add explicit ids in the flatbuffer file.  This will help with merge
conflicts going forwards, and these fields really can't change
anymore...

We want to run the camera application once per pi.

gflags and glog look at argv[0] to determine the main module for
printing help information out.  argv[0] is whatever the binary is
called.

AOS also looks at argv[0] to determine which application is being
started, to name the event loop accordingly, and to find the Application
entry corresponding to the application being run.

We don't care about gflags/glog when running, so we could easily have 1
application entry (with a different name) per node.  That makes it much
harder though to run applications by hand and have them pick up
everything from the config correctly.  One would need to set argv[0], or
symlink the application, or move it to pick up the right name.  Nobody
is going to remember to do that.

This really says that we want to be able to run the same application on
each pi, which means the node list needs to be a list, not an entry.
Luckily, we have no logs we care about with it as a list, so we can just
change it.

Change-Id: I24292c0f54a531e63f5af5e08c613de160278cb8