Add support for tags in a config
These make it easier to support different combinations of nodes by
assigning roles and other information to each node.

This needs to be in the config, since if we want to remap, we need to
use information in the log file header to remap rather than information
in any messages.

Change-Id: I0bc34c5fa906939becaaa0f42f98809eba44ca0f