Regenerate //debian:emscripten_toolchain.patch
For some reason the newer bazel version fails to apply the emscripten
toolchain patch. Maybe they changed their patch tool or something.
Either way, I re-generated the patch here so that newer bazel applies
it without complaints.

The order of files within the patch is different, but it looks to be
functionally equivalent to what was there before.

Change-Id: I578ae6845b1074966cc71391fac3400102ae160f