Teach PartsMessageReader to verify out of order
We want to know when a log file doesn't meet our requirements.
PartsMessageReader is a reasonable spot since it knows a lot about log
files.

Add a test while we are here, and also make it so we get max_time back
for the newest timestamp when the log file finishes to make it easy to
read until X happens.

Change-Id: I934f0dffedf023e10596de558dedc79bdc289ff2