Make messages sent between RunFor call watchers in sim
The prior behavior was that watchers only got called for messages sent
while running.  This caught a lot of people off guard for no good
reason and interacted poorly with simulated_message_gateway.

There is still a small edge case where you:

1) run
2) send a message on event loop 2 on channel 1
3) make a watcher on event loop 1 on channel 1
4) run

It is a bit hard to define what "correct" is here.  Instead, just
prevent the user from writing that for now until we get a better
requirement.  The analog for the real world is a bit wonky since if you
did this with a ShmEventLoop, you'd be quite not realtime and probably
cause problems with watchers.

With all that in mind, the following now works:

1) make a watcher on event loop 1 on channel 1
2) run
3) send a message on event loop 2 on channel 1
4) run

The watcher gets called now.

Change-Id: I75ffb83f5967750ce9565011ba6b3a8b7bafb196