Add basic autonomous program.
The robot is in autonomous mode in the beginning of the match. This code makes it simply drive forward in autonomous mode.
It takes a command line argument for either executing spline code or driving forward.
Removed c++ 17 initialzation with curly braces because the roborio compiler is not compatible with it.

Change-Id: Ia80fb5ce8bba219cbbf4fc46e99ffdbe36cd0d93