Add TimestampMapper to match timestamps with data
This finishes the main sorting code for
https://docs.google.com/document/d/1RZ6ZlADRUHmwiFOOmXA87FHPLFuN-7mS7tbFCwguZDE/edit#

This creates an object to buffer sorted data in, and to buffer data to
match every time we find a timestamp.  Currently, there is no timing out
of the data buffered, but only the data that *could* be forwarded is
buffered.

We also have a bunch of restrictions here to simplify the logic.  The
plan is to relax them as we run into them rather than solve everything
at the start.  And, we can then add tests a lot better.

Change-Id: Idbc515e5594bc031139c7b994aaa71826ff68c0a