Switch LogReader to the new API, and remove the old
Now that we have a fancy new log sorter, lets flip over to using it.
This doesn't simplify anything yet on the log reader side to take
advantage of the simpler code.

While we are here, the new API really wants a LogFiles object instead of
vectors of strings.  Convert any calls over to the new LogFiles API.

The log file was updated to add UUIDs.  They are required with
multi-node log files, and it doesn't seem worth changing that
requirement for this old log.

Change-Id: I84bd63c7339ec43ed01c106131153e1cb6d213bb