Support integrating third_party event loops
Add ForgetClosedFd and BeforeWait to support third_party event loops.
When a file descriptor is closed, it is automatically deleted from
epoll.  This necesitates a ForgetClosedFd function to remove it.

Lots of third party implementations support integration with select
loops.  The interface is that they provide a method to call before going
to sleep which lists out any file descriptors to handle.  This
necesitates a callback to be called before going to sleep.

Change-Id: Id7754785e32322c0a8f57ca7f5e865c05f26589a