Handle log files with only one direction of data
We have seen some log files where A -> B works, but B -> A hasn't
started up yet when interesting things are happening in the log.  There
is enough information to order events causally between nodes, though not
well.  So, use this information to fit what we have.

This doesn't yet handle transitioning between not having both directions
and having both directions.  Both of those behaviors should result in a
jump in time and therefore a crash.

Change-Id: I684c1b61eb18efe2caaedc0e1e86f147c1e30f70