Use config found in SortLogs instead of random log files
This both gives us a single Configuration object to use for all
matching configuration, and also gives us an abstraction to use when we
move the config out of the top of each log file and into a separate
file.

While we are here, expose name from the log file header as well.  This
removes the last user of the raw LogFileHeader from logger.

Change-Id: I0c99d64f9a7222e17100650cdf4b018ae224887a